Rename chcp.py to win_chcp.py
